Move prop-alias-resolution.ts to react helpers
