+++ packages/analysis/src/jsts/rules/S6767/decorator-indirect-prop-usage.ts
@@ -33,7 +33,7 @@ import { areSameTypeDeclarations, getTypeFromTreeNode } from '../helpers/type.js
 import {
   isNamedPropExpressionOrAlias,
   isWholePropsExpressionOrAlias,
-} from './helpers/prop-alias-resolution.js';
+} from '../helpers/react/prop-alias-resolution.js';
 import { isSupportedWholePropsUsage } from './whole-props-usage.js';
 
 /**
